fix: rename directory and namespace in login context
